Fix code inspection for kernel module (#31209)

+++ mode/type/cluster/repository/provider/zookeeper/src/main/java/org/apache/shardingsphere/mode/repository/cluster/zookeeper/listener/SessionConnectionListener.java
@@ -39,7 +39,7 @@ import java.util.Properties;
 @Slf4j
 public final class SessionConnectionListener implements ConnectionStateListener {
     
-    private static final int RECONNECT_INTERVAL_SECONDS = 5;
+    private static final long RECONNECT_INTERVAL_SECONDS = 5L;
     
     private final InstanceContext instanceContext;
